Add files via upload
An update to the main code. I simplified the program by converting repeating code segments and setting them as functions. There is definitely more refinement possible but it's a great start already :D

At the same time, I have also added the map bot. This just takes input, such as 00RRB0 and outputs all the lines in the appropriate format to be used by the main bot (0,0,R,R,B,0).